front end finished

+++ app/Http/Controllers/TagsController.php
@@ -107,6 +107,11 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+        if($tag->posts->count()> 0){
+            session()->flash('error', 'Tags Can\'t be deleted because have some posts.');
+            return redirect()->back();
+        }
+
         $tag->delete();
         session()->flash('success', 'Tag Deleted successfully.');
